Move 'llm' package inside 'components'
