+++ test/conftest.py
@@ -12,7 +12,7 @@ except Exception:
 
 import requests
 import pytest
-from vectoria_lib.llm.llm_factory import LLMFactory
+from vectoria_lib.components.llm.llm_factory import LLMFactory
 from vectoria_lib.common.paths import TEST_DIR
 from vectoria_lib.common.config import Config
 from vectoria_lib.common.io.file_io import get_file_io
